backend/requirements.txt
# FastAPI and Server
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.6

# Database
supabase==2.3.4
sqlalchemy==2.0.25
asyncpg==0.29.0

# Google APIs
google-auth==2.28.1
google-auth-oauthlib==1.2.0
google-api-python-client==2.118.0

# OpenAI
openai==1.12.0

# Utilities
python-dotenv==1.0.1
pytz==2024.1
httpx==0.26.0
pydantic==2.6.1
pydantic-settings==2.1.0
apscheduler==3.10.4